[DEV 2.0] Add new last_updated field
This allows us to sort the files in the file
panel by more recently updated items on the
top.

Consequently, added a mechanism that
updates the field to dt.now() (datetime.datetime.now()), every time a POST/PATCH is
done.

The GET sorts the session objects by descending order of when they were last updated.